Created structure in the project * Moved helpers to separate folder, so it's clear the 3cqsbot.py is the only executable py file * Moved config to separate folder, so it's possible to load it in a docker image from external volume * Updated README.MD * Added sample value in some of the config values, so script doesn't break at first run
